refactor linting errors fixed

+++ contact/views.py
@@ -1,13 +1,11 @@
 from django.shortcuts import render, get_object_or_404
 from django.contrib import messages
-from django.http import HttpResponse
 from django.contrib.auth.decorators import login_required
 from django.core.mail import send_mail
 from .models import Message
 
 from .forms import ContactForm
 
-# Create your views here.
 
 def contact(request):
     """ A view to return the Contact page """
@@ -19,15 +17,19 @@ def contact(request):
     else:
         form = ContactForm()
 
-
     context = {
         'form': form,
     }
     return render(request, 'contact/contact.html', context)
 
 
-def messages(request):
+@login_required()
+def view_messages(request):
     """ A view to return the Messages page """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you must be admin to view user messages.')
+        return redirect(reverse('home'))
 
     user_messages = Message.objects.all()
 
@@ -38,9 +40,12 @@ def messages(request):
     return render(request, 'contact/view_messages.html', context)
 
 
+@login_required()
 def delete_message(request, message_id):
     ''' delete user messages '''
-    print(message_id)
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you must be admin to delete user messages.')
+        return redirect(reverse('home'))
     message = get_object_or_404(Message, pk=message_id)
     message.delete()
     user_messages = Message.objects.all()
@@ -51,8 +56,12 @@ def delete_message(request, message_id):
     return render(request, 'contact/view_messages.html', context)
 
 
+@login_required()
 def reply_to_message(request, message_id):
     """ Allows admin to reply to user messages """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you must be admin to add reply to user messages.')
+        return redirect(reverse('home'))
     message = get_object_or_404(Message, pk=message_id)
 
     context = {
@@ -61,8 +70,13 @@ def reply_to_message(request, message_id):
     return render(request, 'contact/reply.html', context)
 
 
+@login_required()
 def send_email_message(request, message_id):
     """ Allows admin to reply to messages """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you must be admin to add products.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         # Get data from form
         email = request.POST.get['to_email']
